Refactor application code into http directory (#16)
* update composer.json

* move all domain code into their own directory
