+++ App/Http/Domain/Admin/Accounts/Actions/CreateAccountAction.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 
 
-namespace App\Actions\Account;
+namespace App\Http\Domain\Admin\Accounts\Actions;
 
 
 use App\Models\Admin\Account;
@@ -14,23 +14,27 @@ use App\Src\Session\Session;
 use App\Src\State\State;
 use App\Src\Translation\Translation;
 
-final class UpdateAccountDataAction extends Action
+class CreateAccountAction extends Action
 {
     private Account $account;
-    private User $user;
     private Session $session;
 
     protected string $name;
+    protected string $email;
+    protected string $password;
+    protected string $confirmationPassword;
     protected int $rights;
 
     public function __construct(Account $account)
     {
         $this->account = $account;
-        $this->user = new User();
         $this->session = new Session();
         $request = new Request();
 
         $this->name = $request->post('name');
+        $this->email = $request->post('email');
+        $this->password = $request->post('password');
+        $this->confirmationPassword = $request->post('confirmationPassword');
         $this->rights = (int) $request->post('rights');
     }
 
@@ -39,14 +43,19 @@ final class UpdateAccountDataAction extends Action
      */
     protected function handle(): bool
     {
-        $this->account->update($this->account->getID(), [
+        $this->account->create([
             'account_name' => $this->name,
+            'account_email' => $this->email,
+            'account_password' => (string) password_hash(
+                $this->password,
+                PASSWORD_ARGON2ID
+            ),
             'account_rights' => (string) $this->rights,
         ]);
 
         $this->session->flash(
             State::SUCCESSFUL,
-            Translation::get('admin_edited_account_successful_message')
+            Translation::get('admin_create_account_successful_message')
         );
         return true;
     }
@@ -60,17 +69,6 @@ final class UpdateAccountDataAction extends Action
             return false;
         }
 
-        if ($this->account->getID() === $this->user->getID()
-            && $this->rights !== $this->user->getRights()
-        ) {
-            $this->session->flash(
-                State::FAILED,
-                Translation::get('cannot_edit_own_account_rights_message')
-            );
-
-            return false;
-        }
-
         return true;
     }
 
@@ -80,11 +78,35 @@ final class UpdateAccountDataAction extends Action
     protected function validate(): bool
     {
         if (empty($this->name)
+            || empty($this->email)
+            || empty($this->password)
+            || empty($this->confirmationPassword)
             || empty($this->rights)
         ) {
             $this->session->flash(
                 State::FAILED,
                 Translation::get('form_message_for_required_fields')
+            );
+
+            return false;
+        }
+
+        if (!filter_var($this->email, FILTER_VALIDATE_EMAIL)) {
+            $this->session->flash(
+                State::FAILED,
+                sprintf(
+                    Translation::get('form_invalid_email_message'),
+                    $this->email
+                )
+            );
+
+            return false;
+        }
+
+        if ($this->password !== $this->confirmationPassword) {
+            $this->session->flash(
+                State::FAILED,
+                Translation::get('admin_passwords_are_not_the_same_message')
             );
 
             return false;
@@ -97,6 +119,18 @@ final class UpdateAccountDataAction extends Action
             $this->session->flash(
                 State::FAILED,
                 Translation::get('admin_invalid_rights_message')
+            );
+
+            return false;
+        }
+
+        if (!empty((array) $this->account->getByEmail($this->email))) {
+            $this->session->flash(
+                State::FAILED,
+                sprintf(
+                    Translation::get('admin_email_already_exists_message'),
+                    $this->email
+                )
             );
 
             return false;
