+++ App/Http/Domain/Admin/Accounts/Actions/UpdateAccountEmailAction.php
@@ -2,11 +2,10 @@
 declare(strict_types=1);
 
 
-namespace App\Actions\Account;
+namespace App\Http\Domain\Admin\Accounts\Actions;
 
 
 use App\Models\Admin\Account;
-use App\Models\User;
 use App\Src\Action\Action;
 use App\Src\Core\Request;
 use App\Src\Security\CSRF;
@@ -14,16 +13,12 @@ use App\Src\Session\Session;
 use App\Src\State\State;
 use App\Src\Translation\Translation;
 
-class CreateAccountAction extends Action
+final class UpdateAccountEmailAction extends Action
 {
     private Account $account;
     private Session $session;
 
-    protected string $name;
     protected string $email;
-    protected string $password;
-    protected string $confirmationPassword;
-    protected int $rights;
 
     public function __construct(Account $account)
     {
@@ -31,11 +26,7 @@ class CreateAccountAction extends Action
         $this->session = new Session();
         $request = new Request();
 
-        $this->name = $request->post('name');
         $this->email = $request->post('email');
-        $this->password = $request->post('password');
-        $this->confirmationPassword = $request->post('confirmationPassword');
-        $this->rights = (int) $request->post('rights');
     }
 
     /**
@@ -43,19 +34,13 @@ class CreateAccountAction extends Action
      */
     protected function handle(): bool
     {
-        $this->account->create([
-            'account_name' => $this->name,
+        $this->account->update($this->account->getID(), [
             'account_email' => $this->email,
-            'account_password' => (string) password_hash(
-                $this->password,
-                PASSWORD_ARGON2ID
-            ),
-            'account_rights' => (string) $this->rights,
         ]);
 
         $this->session->flash(
             State::SUCCESSFUL,
-            Translation::get('admin_create_account_successful_message')
+            Translation::get('admin_edited_account_successful_message')
         );
         return true;
     }
@@ -77,12 +62,7 @@ class CreateAccountAction extends Action
      */
     protected function validate(): bool
     {
-        if (empty($this->name)
-            || empty($this->email)
-            || empty($this->password)
-            || empty($this->confirmationPassword)
-            || empty($this->rights)
-        ) {
+        if (empty($this->email)) {
             $this->session->flash(
                 State::FAILED,
                 Translation::get('form_message_for_required_fields')
@@ -98,27 +78,6 @@ class CreateAccountAction extends Action
                     Translation::get('form_invalid_email_message'),
                     $this->email
                 )
-            );
-
-            return false;
-        }
-
-        if ($this->password !== $this->confirmationPassword) {
-            $this->session->flash(
-                State::FAILED,
-                Translation::get('admin_passwords_are_not_the_same_message')
-            );
-
-            return false;
-        }
-
-        if ($this->rights !== User::ADMIN
-            && $this->rights !== User::SUPER_ADMIN
-            && $this->rights !== User::DEVELOPER
-        ) {
-            $this->session->flash(
-                State::FAILED,
-                Translation::get('admin_invalid_rights_message')
             );
 
             return false;
