+++ App/Http/Domain/Admin/Accounts/User/Actions/UpdateUserPasswordAction.php
@@ -2,10 +2,10 @@
 declare(strict_types=1);
 
 
-namespace App\Actions\Account\User;
+namespace App\Http\Domain\Admin\Accounts\User\Actions;
 
 
-use App\Models\Admin\Account\Account;
+use App\Models\Admin\Account;
 use App\Models\User;
 use App\Src\Action\Action;
 use App\Src\Core\Request;
